refactor: mv entity org tabs into page content component
